This update is chock full of updates, including: Users are no longer able to move items in the /bf tourney menu Tournament Creation through commands now supports decimals for the time Removed a bug that allowed any item to be turned into a fish Tournaments no longer strip rewards of custom data such as enchantments or names There is also a new mode that can be enabled through the config called Tournament Only. With this mode custom fish will only appear while a tournament is actively running. Otherwise players will only receive basic fish.
To enable Tournament Only Mode, you need to have
Tournament Only: true
in your config file. All new config files that are generated will have the option set to false.

+++ src/com/kunfury/blepFishing/Commands/CommandHandler.java
@@ -127,7 +127,7 @@ public class CommandHandler implements CommandExecutor {
     	if(args[0].equalsIgnoreCase("StartTourney")) {
     		if(sender.hasPermission("bf.admin")) {
     			try {
-        			new Tournament().CreateTourny(sender, args[1], Integer.parseInt(args[2]), Integer.parseInt(args[3]), 
+        			new Tournament().CreateTourny(sender, args[1], Double.parseDouble(args[2]), Integer.parseInt(args[3]),
         					args[4], Integer.parseInt(args[5]));
         			//Oh boi why did I ever write it like this
         			//This is impractical as all hell
